+++ lib/fform.dart
@@ -1,11 +1,16 @@
 library;
 
-export 'src/fform.dart';
-export 'src/fform_exception.dart';
-export 'src/fform_field.dart';
-export 'src/fform_observer.dart';
-export 'src/mixins/cached_validator.dart';
+export 'src/core/fform.dart';
+export 'src/core/fform_exception.dart';
+export 'src/core/fform_field.dart';
+export 'src/core/fform_observer.dart';
+
+// Mixins for FFormField
+export 'src/mixins/cached_validator_field.dart';
+export 'src/mixins/focus_field.dart';
 export 'src/mixins/keyed_field.dart';
-export 'src/mixins/mixin_async_validator.dart';
+export 'src/mixins/mixin_async_validator_field.dart';
+
+// Widgets
 export 'src/widgets/fform_builder.dart';
 export 'src/widgets/fform_provider.dart';
