+++ lib/src/core/fform_exception.dart
